#18979 move and rename endpoint
