+++ dotCMS/src/main/java/com/dotcms/rest/RoleResource.java
@@ -1,11 +1,9 @@
 package com.dotcms.rest;
 
-import javax.ws.rs.DefaultValue;
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
-import javax.ws.rs.QueryParam;
 import javax.ws.rs.core.CacheControl;
 import javax.ws.rs.core.Context;
 import javax.ws.rs.core.Response;
@@ -13,34 +11,23 @@ import com.dotcms.rest.exception.ForbiddenException;
 import com.dotmarketing.business.APILocator;
 import com.dotmarketing.business.Role;
 import com.dotmarketing.business.RoleAPI;
-import com.dotmarketing.business.UserAPI;
-import com.dotmarketing.exception.DoesNotExistException;
 import com.dotmarketing.exception.DotDataException;
 import com.dotmarketing.exception.DotSecurityException;
-import com.dotmarketing.util.Logger;
 import com.dotmarketing.util.UtilMethods;
 import com.dotmarketing.util.json.JSONArray;
 import com.dotmarketing.util.json.JSONException;
 import com.dotmarketing.util.json.JSONObject;
-import com.liferay.portal.model.User;
-import com.liferay.util.StringPool;
-
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.stream.Collectors;
 
 
 @Path("/role")
 public class RoleResource {
 
     private final WebResource webResource = new WebResource();
-    private final RoleAPI     roleAPI     = APILocator.getRoleAPI();
-	private final UserAPI     userAPI     = APILocator.getUserAPI();
 
     /**
 	 * <p>Returns a JSON representation of the Role with the given id, including its first level children.
@@ -377,72 +364,6 @@ public class RoleResource {
 		return jsonChildren;
 	}
 
-	/**
-	 * Load the user and roles by role id.
-	 * @param request   {@link HttpServletRequest}
-	 * @param response  {@link HttpServletResponse}
-	 * @param roleId    {@link String} role
-	 * @param roleHierarchyForAssign {@link Boolean} true if want to include the hierarchy, false by default
-	 * @param roleNameToFilter {@link String} prefix role name, if you want to filter the results
-	 * @return Response
-	 * @throws DotDataException
-	 * @throws DotSecurityException
-	 */
-	@GET
-	@Path("/users/id/{roleid}")
-	@Produces("application/json")
-	@SuppressWarnings("unchecked")
-	public Response loadUsersAndRolesByRoleId(@Context final HttpServletRequest request,
-								   @Context final HttpServletResponse response,
-								   @PathParam   ("roleid") final String roleId,
-								   @DefaultValue("false") @QueryParam("roleHierarchyForAssign") final boolean roleHierarchyForAssign,
-								   @QueryParam  ("name") final String roleNameToFilter) throws DotDataException, DotSecurityException {
 
-		new WebResource.InitBuilder(this.webResource).requiredBackendUser(true)
-				.requiredFrontendUser(false).requestAndResponse(request, response)
-				.rejectWhenNoUser(true).init();
-
-		final Role role = this.roleAPI.loadRoleById(roleId);
-
-		if (null == role || !UtilMethods.isSet(role.getId())) {
-
-			throw new DoesNotExistException("The role: " + roleId + " does not exists");
-		}
-
-		final List<Role> roleList = new ArrayList<>();
-		final List<User> userList = new ArrayList<>();
-
-		Logger.debug(this, ()->"loading users and roles by role: " + roleId);
-
-		if (!role.isUser()) {
-
-			userList.addAll(this.roleAPI.findUsersForRole(role, roleHierarchyForAssign));
-			roleList.addAll(roleHierarchyForAssign? this.roleAPI.findRoleHierarchy(role): Arrays.asList(role));
-		} else {
-
-			userList.add(this.userAPI.loadUserById(role.getRoleKey(), APILocator.systemUser(), false));
-		}
-
-		for (final User user : userList) {
-
-			final Role roleToTest = this.roleAPI.getUserRole(user);
-			if (roleToTest != null && UtilMethods.isSet(roleToTest.getId())) {
-
-				roleList.add(roleToTest);
-			}
-		}
-
-		return Response.ok(new ResponseEntityView(
-				null != roleNameToFilter? this.filterRoleList(roleNameToFilter, roleList):roleList)).build();
-	}
-
-	private final List<Role> filterRoleList(final String roleNameToFilter, final List<Role> roleList) {
-
-		final String roleNameToFilterClean = roleNameToFilter.toLowerCase().replaceAll( "\\*", StringPool.BLANK);
-		return UtilMethods.isSet(roleNameToFilterClean)?
-				roleList.stream().filter(myRole -> myRole.getName().toLowerCase()
-					.startsWith(roleNameToFilterClean)).collect(Collectors.toList()):
-				roleList;
-	}
 
 }
